backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0

# Auth
pyjwt[crypto]==2.9.0

# Validation
pydantic[email]==2.9.0
pydantic-settings==2.5.0

# Data processing
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1

# ML
scikit-learn==1.5.2
statsmodels==0.14.4
prophet==1.1.6
tensorflow==2.16.2
xgboost==2.1.2
joblib==1.4.2

# Admin
sqladmin==0.19.0

# WebSocket
websockets==13.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
